Refactor memory to its own module
